+++ cmd/version.go
@@ -24,7 +24,7 @@ import (
 
 // ReleaseVersion is the release version of the codebase.
 // Usually overridden by tag names when building binaries.
-var ReleaseVersion = "local build (latest release 1.10.0-pre)"
+var ReleaseVersion = "local build (latest release 1.10.0)"
 
 // versionCmd represents the version command
 var versionCmd = &cobra.Command{
