+++ cmd/wallet/sharedexport/input.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package walletsssexport
+package walletsharedexport
 
 import (
 	"context"
